Refactored, moving common datastore logic into st2common/services/datastore.py. Not sure if that's the appropriate place.
